+++ pybillomat/__init__.py
@@ -5,5 +5,5 @@ Python-Billomat - Billomat API Client Library
 """
 
 from http import Connection
-from clients import Client, Clients
+from clients import Client, Clients, ClientsIterator
 from invoices import Invoice, Invoices